requirements.txt
aiopg
websockets
aiohttp
discord.py
audioop-lts
pyxivapi~=0.5.1
python-dateutil
pytz
guess-language
